be more careful about how we re-open the repository
Summary:
Update the hg_import_helper.py code to catch exceptions from `repo.close()`
when we are re-opening the repository after an error.  Also set `self.repo` to
`None` before we start, so that it will be left as None if anything goes wrong
either closing or re-opening the repository.

I ran into a situation where `repo.close()` threw an error, and previously
this would leave `hg_import_helper.py` stuck in a bad state since it would
still be pointing to the old repo object.  The next time it received a command
it would fail and try to call `self.repo.close()` again, which would still
fail since the repository was already in a halfway closed state.

Now the code will always forget about the old repository object and create a
new object.

Reviewed By: wez

Differential Revision: D9419109

fbshipit-source-id: 15bb296ba19d9d3d2a2b90169bf25b0e8e197c1f