don't allow state-enter/leave event errors to block hg
Summary:
In these threads, our users are working around what is
either an invalid state transition on the hg side and/or slow
processing of the state transitions on the watchman side.

Since the state updates are purely advisory, we do not want to
block users in this case.

This doesn't address the root cause of this problem, it just stops
it from breaking our users.

Reviewed By: quark-zju

Differential Revision: D9437674

fbshipit-source-id: a2b0514a71e31ee7c92df1c24031b12a349822b5