update `eden debug overlay` to support extracting contents
Summary:
Add an `--extract-to` flag to `eden debug overlay` to ask it to copy inode
data out of the overlay.  When used with a file inode this copies the file
contents.  When used with a directory inode this copies the directory
structure and file contents for all materialized files under that directory.

This makes it easier to extract user's untracked and modified files out of the
overlay when recovering state after an unclean filesystem shutdown.

Reviewed By: wez

Differential Revision: D9445559

fbshipit-source-id: 9abcf82695014d79f248ba5fdb09727a49f3098a