extutil: make runbgcommand Ctrl+C safe
Summary:
There is a user report that a command was strangely forked into 2 processes:

The blackbox looks like:

  .... (no other "commit" command around this time)
  2018/08/21 10:57:03 c00b (33973)> commit
  2018/08/21 10:58:59 c00b (34191)> sl
  2018/08/21 10:58:59 c00b (34159)> pythonhook-pretxncommit: hgext.checkmessagehook.checkcommitmessage finished in 0.00 seconds
  2018/08/21 10:58:59 7b7a (34159)> commit exited 0 after 116.00 seconds
  2018/08/21 10:58:59 c00b (34191)> sl exited 0 after 0.39 seconds
  2018/08/21 10:58:59 7b7a (33973)> commit exited -1 after 116.17 seconds

It appears that pid 34159 was a fork of pid 33973. The fork and the original
process both end and mess up with the transaction framework.

The user did pressed Ctrl+C.  Looking at the code, I suspect extutil suffers
a same issue that I fixed for upstream worker.py [1]. So let's fix extutil
too.

[1]: https://www.mercurial-scm.org/repo/hg/rev/31763785094bb1bf604245fa22a7d9960662477c

Reviewed By: singhsrb

Differential Revision: D9438093

fbshipit-source-id: 66d72d5d1d2be89f078f391eac9bd7b21b4c176c