remotenames: fix a crash in remotenames() revset evaluation
Summary:
Update the `remotenames()` revset to check if the repository in question
actually supports remotenames.  Otherwise the code will crash trying to access
`repo._remotenames` if it does not exist.

The `_findrecenttree()` function in the `treemanifest` extension evaluates the
`remotenames()` revset if the remotenames extension is loaded.  This can cause
it to evaluate this revset even if the repository in question was initialized
before the remotenames extension was loaded, and therefore does not support
remotenames.

Reviewed By: quark-zju, strager

Differential Revision: D9486916

fbshipit-source-id: 4a656f5f56330b4038afdb2f88e35b7c97bbc81e