diff allow remote obsoleted revisions to be pulled if missing
Summary:
The goal is to improve backup experience, make repo after recovery to behave as it was for the most popular commands.

This is only if backup is backed by cloud sync because we have more meta information in the repo.

the commands:

`hg diff --since-last-submit`
`hg log -r 'lastsubmitted(.)'`
`hg unamend`

for example, are not working for fresh repos that are recloned because the repo will only contains the latests snapshot of commits and also all obsmarker history but not obsoleted commits.

those are popular commands, so let's make them work.

the standard error there is 'unknown revision' and it is very confusing for users

there is no reason not to pull a remote obsoleted revision into the repo if needed, and there is no reason to ask user to do that manually.

the obstore is still lazy loaded, so shouldn't be perf problems

this diff supports good range of commands, so should make experience for our backup use case users significantly better.

pulling the hidden revisions is also relatively fast and doesn't affect commit cloud state of the repo in any way

Reviewed By: quark-zju

Differential Revision: D9437518

fbshipit-source-id: 01065c642aa9a194f2d321f03c1bd747f57c74b1