update the test HgRepo code to preserve HG_REAL_BIN and CHG_BIN
Summary:
In D9486915 I updated the code to invoke `hg` from the local repository, but
this was invoking the telemetry wrapper, and this was still falling through
and running the system-installed `hg.real` binary rather than the from the
local repository.  This fixes the code to also pass through the `HG_REAL_BIN`
and `CHG_BIN` environment variables when invoking the `hg` wrapper so that the
underling real hg code is picked up from the current repository as wel.

Reviewed By: chadaustin

Differential Revision: D9494271

fbshipit-source-id: 89c5154f501536b9d7a43d472b1204f10e5dc664