perform additional verification on CMD_CAT_FILE responses
Summary:
This updates the `hg_import_helper.py` code so that when importing files it
also includes the file length in the response, in addition to the response
contents itself.  This information is redundant and exists purely to help
the receiving C++ code verify that the response is in fact completely valid.

This is implemented with a new command number, and `hg_import_helper.py` still
supports the old command number with the old behavior so that it will still
work if old edenfs daemon instances start new instances of
`hg_import_helper.py`.

Reviewed By: chadaustin

Differential Revision: D9494653

fbshipit-source-id: 34d1221b6048d6016b30fd7f606ca0b140552d80