add a wrapper class around HgImporter to allow retrying
Summary:
Add a wrapper class around HgImporter that can catch errors communicating with
the underlying `hg_import_helper.py` script.  If something goes wrong
communicating with the python code we restart the import helper and retry
the operation once.

This code only retries on HgImporterError exceptions, and nothing throws these
at the moment.  Therefore for now this should not result in any behavior
changes.  In the future I will update our error handling to throw
HgImporterError when we fail to communicate with the underlying
hg_import_helper.py script.

Reviewed By: wez

Differential Revision: D9507291

fbshipit-source-id: 0167959df6292a577942e049576669fb6711ce76