run-tests: truncate diff output
Summary:
`.t` tests are stateful. This means if some early parts do not execute
as the test expect, the the remaining parts can break in funny ways,
and generate lots of uninteresting output.

Limit the length of the diff output so the "likely" boring parts won't show
up by default. The change can be overridden by `--maxdifflines`.

Reviewed By: phillco

Differential Revision: D9516041

fbshipit-source-id: 6c998cdc73ae4a45e2f4167c22c04df51392da29