add tests for HgImporter error handling logic
Summary:
Add a new unit test that exercises HgImporter and HgBackingStore using an
alternate hg_import_helper.py script that can inject errors into the response
data stream.  This tests the behavior of HgImporter when there are
serialization errors with the data, as opposed to legitimately-serialized
error responses containing a python exception message.

Reviewed By: wez

Differential Revision: D9510568

fbshipit-source-id: ae86f69d5559701e9e8a9a18569986d64488eaf4