hg: move to more fine-grained python execution from the binary
Summary:
This steals lots of code from [1] and uses dev versions of crates from [2]
as well.

This approach is better than calling things via `Py_Main` directly because
it allows us to pre-populate the container with some data (pre-population
needs to happen after `Py_Initialize` and before passing control to
current Python-based hg).

NB: returning `1` when the exception instance does not have a `code` attribute is an experimentally-determined decision. With `255` tests fail, so `1` just matches Python behavior. I guess we can change it anytime if we come up with exit code globbing for the tests.

Reviewed By: quark-zju

Differential Revision: D9482436

fbshipit-source-id: 9ca63f2358ddc1de60228c489aa9d02b4b7bc482