increase the timeouts in HgPrefetchTest
Summary:
This test seems to fail frequently while waiting on the tree data for the
"src" tree (at line 182).  The code previously was using a 1s timeout for this
operation; this changes it to use a 10s timeout.

I'm not sure what would have changed that makes this take longer now.  It's
possible something changed with the treemanifest code on the mercurial side.
It is somewhat surprising to me that this would regularly take more than 1s
though since this is only computing data locally about a pretty small test
repository.

Reviewed By: wez

Differential Revision: D9583645

fbshipit-source-id: 24695cc1b940540ec32461e2d80c30a3fd87a3e3