add an --overlay flag to the fsck command
Summary:
Add an --overlay flag to the `eden fsck` command to allow running it on an
alternative overlay directory.

This is similar to the `--overlay` flag added to the `eden debug overlay`
command in D9445560.

At the moment a checkout path is still required when running
with `--overlay`.  This is used to find the socket to talk to edenfs.  This is
not actually used at the moment, but in the future we may need this in order
to help fix some of the filesystem state in some cases.  (For instance to get
the current blob or tree ID for a path in the current commit.)

Reviewed By: wez

Differential Revision: D9615334

fbshipit-source-id: 5a0da55f00429f596be5d86e5303e1fcdeff9ea7