improve some of the fsck output
Summary:
Wait to print errors until after we have finished scanning all of the overlay
files.  Previously we printed errors as they were detected.  However we would
not be able to print the path names correctly if we had not processed all of
the inode's parents yet.  This resulted in error messages incorrectly
displaying paths as unlinked in some cases.

Also improve the error strings shown for MissingMaterializedInode and
InvalidMaterializedInode to also include the inode number.  Also correct the
inode type display for InvalidMaterializedInode.

Reviewed By: wez

Differential Revision: D9615336

fbshipit-source-id: eb273d51c937e76ffed0e021da848f5fb940145d