hg: make shelve not abort on every transaction with treemanifest
Summary:
Before this commit, `treemanifest` registered a post-transaction hook to
prefetch data for the parents of new commits. Shelve produces a hidden
commit, so this logic failed, since it used a filtered repo instance.

Let change it to use unfiltered repo and remove all of the hidden commits
from the revset to avoid surprises.

Reviewed By: markbt

Differential Revision: D9613454

fbshipit-source-id: 5caddd2948e5d2e982462bebdb950b359995baa1