hgsubversion: handle corrupted rowcount
Summary:
Handle an empty .rowcount file. This is a common crash in the
field. The root cause is likely the non-atomic write to the .rowcount
file in fileproperty's setter.

The `rowcount` value is only used to accelerate calculating how many
commits pulled for display purpose. So the correctness of this value
does not matter.

Reviewed By: quark-zju

Differential Revision: D9631942

fbshipit-source-id: 2493304bf1cc82b6639100434bae667c2fc1eb70