ensure that logging errors don't mask the original error
Summary:
D9363420 added code to perform extra logging of exceptions.  Unfortunately if
an error occurs inside the logging code (for instance, if it tries to log
binary data that cannot be json-encoded) it masks the original exception
message and backtrace with the logging exception.

This issue is specific to python 2.x.  Python 3.x reports both exception
backtraces in this case.  However, in python 2.x putting a `try...except`
block inside an `except` block ends up overwriting the original exception
info, even when the inner exception is handled.

Moving the inner `try...except` block to its own function is sufficient to
avoid this problem in python 2.x

Reviewed By: singhsrb

Differential Revision: D9764805

fbshipit-source-id: c5a9e65fcd7505ccb478500d62c06c2c94308332