mononoke: fix getfiles timeouts
Summary:
We had a lot of requests that took > 15 mins on Mononoke, while taking few
seconds on mercurial. Turned out that hgcli doesn't play well with big chunks.
Looks like AsyncRead very inefficiently tries to allocate memory, and that
causes huge slowness (T33775046 for more details).

As a short-term fix let's chunk the data on the server. Note that now we have
to make getfiles request streamable and manually insert the size of the
request.

Reviewed By: lukaspiatkowski

Differential Revision: D9738591

fbshipit-source-id: f504cf540bc7d90e2cbebba9808455b6e89c92c6