always unload trees prior to checkout
Summary:
The checkout operation is a bit hamstrung, performance-wise, when
inodes are loaded. If it's possible the kernel has a reference to an
inode, we can't short-circuit the checkout operation and instead must
precisely add and remove children from the source control tree and
persist the new children and inode numbers into the overlay.

It's much faster if we treat checkout as an inode-number-invalidating
change and unload all inodes that aren't referenced by the kernel.

Reviewed By: strager

Differential Revision: D9781532

fbshipit-source-id: 50912cf89dbe9ae1bd7ef91fdf7bf8ee5eda8667