lock: remove stale POSIX symlink lock automatically in `makelock`
Summary:
Previously, removing stale POSIX lock is done at different places:

- Stale flock lock is removed by `posix.makelock`. The "read + check + unlink"
  is done atomically (taking the directory lock).
- Stale (legacy) symlink lock is removed by `lock.py`.
  The "read + check + unlink" is not done atomically, which seems like a scary
  bug [since 2016](https://www.mercurial-scm.org/repo/hg/rev/d314a89fa4f16b7a9b24dc999145d98a878f7ca9).

This patch cleans the logic up so removing stale POSIX lock is done
properly at `util.makelock`.

Note that Windows logic is not changed yet, so it's in theory broken.
Ideally we'd like to use flock-ish locking on Windows too. But that's
outside the scope of this diff.

Reviewed By: ikostia

Differential Revision: D9770935

fbshipit-source-id: fc58f30348f8a9aba22df415c6d2b2dfeb6a87d9