traceprof: always buffer output in a temporary file
Summary:
traceprof uses `fprintf`, which requires a real `FILE *` to write.
That means it does not work for stringio, which is not backed by `FILE *`.

Fix it by always buffering the output in a temporary file.

As we're here, fix the Windows version of traceprof where Python and Cython can
have incompatible `FILE *`s, by only using `FILE *` in the Cython world.

Reviewed By: markbt

Differential Revision: D9828552

fbshipit-source-id: 8271d1aa3eb562ae7669b73af8f80473ef733a68