profiling: always buffer the output
Summary:
The next patch adds a threshold to silence profiling output if the command is
fast. Therefore the profiling output shouldn't be written to terminal or files
directly.

Previously, only blackbox output is buffered. This simplifies the code since
self._fpdoclose, self._closefp, and self._output are no longer necessary.

Reviewed By: markbt

Differential Revision: D9828554

fbshipit-source-id: 7ac45ddc2cd332bd38dd97d08d9f7c3a27f1a16b