profiling: allow profiling config to be specified by other sections
Summary:
Previously, the `[profiling]` config section defines the profiler. Now it can be
defined by other sections, `[profiling:1]`, `[profiling:2]`, etc. Those sections
are sorted by name, and the first one with `enabled = true` is effective.

This makes it possible to specify secondary profiling configs, like having a
low-overhead profiler to be always enabled and logged if the command exceeds
certain threshold:

  [profiling:background]
  enabled = 1
  freq = 1
  minelapsed = 30
  output = blackbox
  type = stat

The config will only be used when the primary profiling (`--profile`, or
`profiling.enabled`) is disabled.

Reviewed By: markbt

Differential Revision: D9828548

fbshipit-source-id: 301a2b049d569d6fc065d05ab5b02e85b4ea8cc5