pushrebase: add debug command to print manifest reads during push
Summary:
Part of push contention debugging. We want to be able to turn this on server-side to see how many uncached manifest reads we're doing.

If `pushrebase.debugprintmanifestreads` is set it'll log to blackbox.

If `pushrebase.debugprintmanifestreads.user` is set it'll also be sent to client logs, which is fine since most pushes are Landcastle these days.

Stolen from my test in D4785464.

Reviewed By: quark-zju

Differential Revision: D9994420

fbshipit-source-id: ff1294dfda2f2774fc6be78e096dad9fd9a28a5e