commitcloud: check if push filter has any effect
Summary:
This is nice for performance reasons

The only new logic is about checking that the revset effectively filter anything out from all set of draft commits.

If not, we just ignore it.

The rest of changes is cosmetic repair.

As we are close to enable everyone I think I will have to set this option

```
[commitcloud]
custom_push_revs=date('>Apr 01 2018')
```

or similar to minimize issues until mononoke and all old commits tree manifest conversion.

Reviewed By: markbt

Differential Revision: D9981311

fbshipit-source-id: fe09ff84f90a692e45ee58170e85b563743736be