pathutil: don't check for case sensitivity on always case-sensitive filesystems
Summary:
When creating a VFS, Mercurial attempts to check if it is case sensitive by
stating the root of the VFS with differently cased names.  If the root of the
VFS is a directory on an autofs mount, this check can be slow.  Don't bother
checking if the VFS root has a filesystem type that we know is always case
sensitive.

Reviewed By: quark-zju

Differential Revision: D9993177

fbshipit-source-id: 4395b16f5f7b32218836422b44dae953dfb8c52b