pushrebase: improve cold manifest logging (track what was cached)
Summary:
Multiple calls to the wrapped function should, in theory, be cached, so that calling them isn't as expensive as the first time.

Let's track what should be cached ourself so we can provide useful output. This is less risky than trying to separate the underlying logic in revlog.py.

Also, properly handle rev numbers (we are called with both) and turn them into nodes.

Reviewed By: quark-zju

Differential Revision: D10023518

fbshipit-source-id: 9a9ac1ed223136ca19d20db2ee6eb1ca11d5ae49